Update graphiql to 0.11.11
The main thing I want from this update is rendering field description as
markdown
https://github.com/graphql/graphiql/commit/6398718c4111360f8c95fda3622bbcfa48cc0862,
but I imagine there's other stuff in the update that might be nice or on
the other hand might be breaking.  Thought I'd put this in to see if you
were open to the update, happy to do any other work needed to get this
through.  I didn't commit the changes from the build script since I
figured that might be part of your release process.